more progressions and expanding functionality of arpeggio sequences
Chord inversions and multi-octave chords are now possible as sequences where indices exceed the number of intervals in the chord. This also allows for mixing tetrads like 7ths with triads. It's unconventional, but it accomplishes so many things in the right way and I can't think of any other way to interpret this usage.